// apps/studio/app/studio/contract-lab/types.ts
export interface AbiEntry {
  type: string;
  name: string;
  inputs?: { name: string; type: string }[];
  outputs?: { type: string }[];
  state_mutability?: string;
  items?: AbiEntry[];
}